requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
password-strength
requests
pytest
python-dotenv
pytest-cov